HTML Skeleton for TA side and Student side
I was thinking that the TA and students see the same waitlist page but the buttons are just present whether or not a TA is logged in. If that is too tricky it would be easier (and not much more code wise) to detect whether a TA is logged in and then map to a different waitlist webpage all together and in that case I could make a StudentWaitlist.html and TAWaitlist.html. Backend peeps let me know asap if you would prefer that